Split off discovery part
